+++ NEWS.md
@@ -1,4 +1,4 @@
-# LWFBrook90R (in development)
+# LWFBrook90R 0.6.3
 
 ## Changes
 * removed default value for `topsoil` arguments in functions `hydpar_hypres()` and `hydpar_hypres_tab()`.